Revise team structures.
Admins shouldn't be in the committers team. Their admin
permissions allow them all the same permissions as the
committers team. It allows repos to split notifications
a bit better. If a person wants to be in both, that can
be controlled in the variables file.

I attempted to clean up the comments a bit more with the
refactor. Some of the comments no longer made sense.

I reduced some of the code with a concat function call
which should help make the repo team definition a bit
easier to understand.