Terraform v2 (#15)
* terraform org mgmt

* remove webhooks

* added permissions to readme

* readme changes

* gha

* Use member as dynamic variable to avoid confusion with the members list.

* Adjust repository defaults.

The wiki seems like a possibility and the security alerts seems
like a good thing to enable by default until someone complains.

* Comment the main terraform file with descriptions of what each section does.

Removes the security template and other sections
that aren't 100% obviously critical.

We can always re-add these later, but teams can also
configure this manually.

* Pin the terraform github action third-party actions to SHA hashes.

* Use domain specific naming for variables.

This should make it easier to determine what a parent team
relates to versus what a child team relates to.

* Re-add review request delegation.

I didn't see it was used for the playground repo.

* Attempt to split the org teams from the repo teams.

* Revise the playbooks for the new names for the variables.

* Fix missed team_parents reference.

* Update terraform specific docs with new variable names.

* The teams are maps, not lists.

These need to be merged rather than concatenated.

* Revert change, use dynamic members.

Adds a comment to hopefully reduce confusion.

* Remove maintainers collections where they are unnecessary.

The Django Commons admin team will have access to all
teams by being owners in the organization. That isn't
controlled by the terraform plan.

* reorg

* Terraform v2

* Revise team structures.

Admins shouldn't be in the committers team. Their admin
permissions allow them all the same permissions as the
committers team. It allows repos to split notifications
a bit better. If a person wants to be in both, that can
be controlled in the variables file.

I attempted to clean up the comments a bit more with the
refactor. Some of the comments no longer made sense.

I reduced some of the code with a setunion function call
which should help make the repo team definition a bit
easier to understand.

* Improve comments of memberships on the repo files.

---------

Co-authored-by: tschilling <schillingt@better-simple.com>